+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-## New versio
+## New version
 - Fix session provisioning timeout and delay handling
+- Fix get_columns_in_relation function to stop returning additional partition columns
 
 ## v1.8.1
 - Fix typo in README.md