+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-## New versio
+## New version
 - Fix session provisioning timeout and delay handling
+- Fix table materialization for Delta models
 
 ## v1.8.1
 - Fix typo in README.md