+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add on_schema_change possibility
 - Fix table materialization for Delta models
 - Change GlueColumn parent from base Column to SparkColumn
+- Fix get_columns_in_relation function to stop returning additional partition columns
 
 ## v1.8.1
 - Fix typo in README.md