+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Change GlueColumn parent from base Column to SparkColumn
 - Fix unraised DbtDatabaseError
 - Fix get_columns_in_relation function to stop returning additional partition columns
+- Fix null values handling in seeds
 
 ## v1.8.1
 - Fix typo in README.md