+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 - Change GlueColumn parent from base Column to SparkColumn
 - Fix unraised DbtDatabaseError
 - Fix get_columns_in_relation function to stop returning additional partition columns
-
+- Fix exceptions import for FailedToConnectError and ExecutableError
+  
 ## v1.8.1
 - Fix typo in README.md
 - Fix unit test failure caused by moto 5 upgrade