+++ CHANGELOG.md
@@ -8,7 +8,8 @@
 - Fix get_columns_in_relation function to stop returning additional partition columns
 - Fix null values handling in seeds
 - Fix exceptions import for FailedToConnectError and ExecutableError
-  
+- Fix the case-sensitive comparison on the seed name
+
 ## v1.8.1
 - Fix typo in README.md
 - Fix unit test failure caused by moto 5 upgrade